long is 4 bytes in wasi :( => use long long where appropriate

+++ source/tests.h
@@ -56,14 +56,14 @@ bool assert_equals_x(Node a, const char *b, char *context = "") {
 
 bool assert_equals_x(Node a, int b, char *context = "") {
 	if (a != Node(b))printf("\nFAILED assert_equals! %d should be %d %s\n"s, a.value.longy, b, context);
-	else printf("OK %ld==%ld\n", a.value.longy, (long) b);
+	else printf("OK %lld==%d\n", a.value.longy, b);
 	return a == b;
 }
 
 // WTF why is char* unexplicitly cast to bool!?!
 bool assert_equals_x(Node a, bool b, char *context = "") {
 	if (a != Node(b))printf("\nFAILED assert_equals! %d should be %d %s\n"s, a.value.longy, b, context);
-	else printf("OK %ld==%ld\n", a.value.longy, (long) b);
+	else printf("OK %lld==%d\n", a.value.longy, b);
 	return a == b;
 }
 
@@ -88,7 +88,7 @@ bool assert_equals_x(Node a, double b, char *context = "") {
 
 bool assert_equals_x(Node a, long b, char *context = "") {
 	if (!(a == b))printf("\nFAILED assert_equals! %s %ld should be %ld %s\n"s, a.name, a.value.longy, b, context);
-	else printf("OK %ld==%ld %s\n", a.value.longy, b, context);
+	else printf("OK %lld==%ld %s\n", a.value.longy, b, context);
 	return a == b;
 }
